Explicitly initialize lastDisplayed
This is a hack to make the time delay logic work. This was not explicit
before, but this makes the hack explicit.

The only time this wouldn't work would be when the server's clock is set
to Jan  1st 1970, in which case you would wrongly wait one frame delay.

Horrible, isn't it.